requirements.txt
discord.py==2.5.2
SQLAlchemy==2.0.41
python-dotenv==1.1.0
aiosqlite==0.19.0
eventlet==0.39.1